+++ spec/factories/performed_segments.rb
@@ -13,11 +13,18 @@ FactoryBot.define do
       # starting_time { Time.new(2017, 2, 1, 15, 0, 0) }
       starting_time { '2015-2-1 15:00:00' }
     end
+
     trait :finlandia do
       category { Category.find_by(name: 'JUNIOR LADIES') }
       segment { Segment.find_by(name: 'FREE SKATING') }
       # starting_time { Time.new(2015, 9, 2, 17, 0, 0) }
       starting_time { '2017-9-1 15:00:00' }
+    end
+
+    trait :ice_dance do
+      category { Category.find_by(name: 'ICE DANCE') }
+      segment { Segment.find_by(name: 'RYTHM DANCE') }
+      starting_time { '2015-2-2 15:00:00' }
     end
   end
 end