+++ spec/factories/scores.rb
@@ -16,8 +16,9 @@ FactoryBot.define do
       result_pdf { 'http://world2017.isu.org/results/men/short.pdf' }
 
       after(:build) do |score|
-        create(:element, :solo_jump, score: score)
+        element = create(:element, :solo_jump, score: score)
         create(:element, :combination_jump, score: score)
+        create(:element_judge_detail, element: element, official: score.performed_segment.officials.first)
         create(:component, :ss, score: score)
       end